+++ packages/builder-util-runtime/src/updateInfo.ts
@@ -42,10 +42,10 @@ export interface UpdateInfo extends VersionInfo {
   /**
    * The release name.
    */
-  readonly releaseName?: string | null
+  releaseName?: string | null
 
   /**
-   * The release notes.
+   * The release notes. List if `updater.fullChangelog` is set to `true`, `string` otherwise.
    */
   releaseNotes?: string | Array<ReleaseNoteInfo> | null