+++ packages/electron-builder/src/cli/cli.ts
@@ -3,11 +3,10 @@
 import { exec, log, warn } from "builder-util"
 import { printErrorAndExit } from "builder-util/out/promise"
 import { cyan, dim, green, reset, underline } from "chalk"
-import { parse as parseEnv } from "dotenv"
-import { readFile, readJson } from "fs-extra-p"
+import { readJson } from "fs-extra-p"
 import isCi from "is-ci"
 import * as path from "path"
-import { orNullIfFileNotExist } from "read-config-file"
+import { loadEnv } from "read-config-file"
 import updateNotifier from "update-notifier"
 import yargs from "yargs"
 import { build, configureBuildCommand } from "../builder"
@@ -77,20 +76,4 @@ async function rebuildAppNativeCode(args: any) {
   await exec(process.platform === "win32" ? "node-gyp.cmd" : "node-gyp", ["rebuild"], {
     env: getGypEnv({version: await getElectronVersion(projectDir), useCustomDist: true}, args.platform, args.arch, true),
   })
-}
-
-async function loadEnv(envFile: string) {
-  const data = await orNullIfFileNotExist(readFile(envFile, "utf8"))
-  if (data == null) {
-    return null
-  }
-
-  const parsed = parseEnv(data)
-  for (const key of Object.keys(parsed)) {
-    if (!process.env.hasOwnProperty(key)) {
-      process.env[key] = parsed[key]
-    }
-  }
-  require("dotenv-expand")(parsed)
-  return parsed
 }