+++ packages/electron-updater/src/Provider.ts
@@ -1,4 +1,4 @@
-import { CancellationToken, HttpExecutor, UpdateInfo, VersionInfo, WindowsUpdateInfo } from "builder-util-runtime"
+import { CancellationToken, HttpExecutor, safeStringifyJson, UpdateInfo, VersionInfo, WindowsUpdateInfo } from "builder-util-runtime"
 import { OutgoingHttpHeaders, RequestOptions } from "http"
 import { URL } from "url"
 import { FileInfo, isUseOldMacProvider } from "./main"
@@ -27,10 +27,10 @@ export abstract class Provider<T extends VersionInfo> {
 
     // noinspection JSDeprecatedSymbols
     if ((info as WindowsUpdateInfo).sha2 == null && info.sha512 == null) {
-      throw new Error(`Update info doesn't contain sha2 or sha512 checksum: ${JSON.stringify(info, null, 2)}`)
+      throw new Error(`Update info doesn't contain nor sha256 neither sha512 checksum: ${safeStringifyJson(info)}`)
     }
     if (info.path == null) {
-      throw new Error(`Update info doesn't contain file path: ${JSON.stringify(info, null, 2)}`)
+      throw new Error(`Update info doesn't contain file path: ${safeStringifyJson(info)}`)
     }
   }