+++ packages/electron-updater/CHANGELOG.md
@@ -1,22 +1,34 @@
-# 2.12.0
+## 2.13.0
+
+### Features
+
+* **electron-updater:** full changelog for all versions from current to latest ([67fe9ff](https://github.com/electron-userland/electron-builder/commit/67fe9ff))
+
+## 2.12.1
+
+### Performance Improvements
+
+* **electron-updater:** a little bit more compact blockmap data ([c92bc38](https://github.com/electron-userland/electron-builder/commit/c92bc38))
+
+## 2.12.0
 
 ### Features
 
 * [Linux auto-update](https://github.com/electron-userland/electron-builder/releases/tag/v19.37.0)
 
-# 2.11.0
+## 2.11.0
 
 ### Features
 
 * Differential updater: use [content defined chunking](https://github.com/electron-userland/electron-builder/releases/tag/v19.36.0)
 
-# 2.10.2
+## 2.10.2
 
 ### Bug Fixes
 
 * Differential updater: fix "To download" in percentage value calculation (cosmetic fix)
 
-# 2.10.1
+## 2.10.1
 
 ### Bug Fixes
 
@@ -35,13 +47,13 @@
 * [Delta updates for Windows Web Installer](https://github.com/electron-userland/electron-builder/releases/tag/v19.28.4).
 
 
-# 2.8.9 (2017-09-01)
+## 2.8.9 (2017-09-01)
 
 ### Bug Fixes
 
 * Electron-updater does not support enterprise Github. [#1903](https://github.com/electron-userland/electron-builder/issues/1903).
 
-# 2.8.8 (2017-09-01)
+## 2.8.8 (2017-09-01)
 
 ### Bug Fixes